move _create_tmp_index_template to conftest
+ autouse and trigger once only during a session
