+++ tests/core/test_views.py
@@ -1,23 +1,7 @@
-import os
 from unittest.mock import patch
 
 import pytest
 from rest_framework.status import HTTP_200_OK
-
-
-@pytest.fixture(autouse=True)
-def _create_tmp_index_template():
-    # setup
-    template_path = "src/core/templates/vue-index.html"
-    if os.path.exists(template_path):
-        yield
-        return
-
-    f = open(template_path, "w")
-    yield
-
-    # teardown
-    os.remove(f.name)
 
 
 @pytest.mark.parametrize(
